StringBuffer replaced by StringBuilder
Reports any variables declared as java.lang.StringBuffer which may be more efficiently declared as java.lang.StringBuilder. java.lang.StringBuilder is a non-thread-safe replacement for java.lang.StringBuffer, available in Java 5 and newer.
This inspection only reports if the project or module is configured to use a language level of 5.0 or higher.